requirements.txt
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0